@@ -8,13 +8,14 @@
     "@chakra-ui/react": "^1.7.4",
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
+    "@react-three/drei": "^9.4.2",
+    "@react-three/fiber": "^8.0.11",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^10.4.9",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^25.2.3",
     "@types/node": "^12.20.42",
     "@types/react": "^16.14.21",
-    "@types/react-dom": "^16.9.14",
     "@types/yup": "^0.29.13",
     "axios": "^0.25.0",
     "dayjs": "^1.11.0",
@@ -24,12 +25,16 @@
     "gh-pages": "^3.2.3",
     "http-status-codes": "^2.2.0",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-div-100vh": "^0.7.0",
+    "react-dom": "^18.0.0",
     "react-icons": "^3.11.0",
     "react-query": "^3.34.12",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "react-scroll-parallax": "^3.0.6",
+    "react-scroll-percentage": "^4.2.0",
+    "three": "^0.139.2",
     "typescript": "^4.5.4",
     "web-vitals": "^0.2.4",
     "yup": "^0.32.11",
@@ -59,6 +64,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-dom": "^18.0.1",
     "@types/uuid": "^8.3.4"
   }
 }
