@@ -1,25 +1,22 @@
-import { ColorModeScript } from '@chakra-ui/react';
 import axios from 'axios';
 import dayjs from 'dayjs';
 import 'dayjs/locale/pl';
 import * as React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 import { App } from './App';
 import reportWebVitals from './reportWebVitals';
 import * as serviceWorker from './serviceWorker';
+import utc from 'dayjs/plugin/utc';
 
 //axios.defaults.baseURL = 'http://localhost:3001/';
 axios.defaults.baseURL = 'https://testnginx.bieda.it/';
-dayjs.locale('pl');
 
-ReactDOM.render(
-  <React.StrictMode>
-    <ColorModeScript />
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+dayjs.locale('pl');
+dayjs.extend(utc);
+
+const root = createRoot(document.getElementById('root') as any);
+root.render(<App />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
