@@ -11,30 +11,39 @@ import { UnauthorizedHandler } from './components/other/UnauthorizedHandler';
 import { Calendar } from './pages/Calendar';
 import { Dashboard } from './pages/Dashboard';
 import { AnimatedTransition } from './components/other/AnimatedTransition';
+import { LandingPage } from './pages/LandingPage';
+import { ParallaxProvider } from 'react-scroll-parallax';
+import './style.css';
+import { ReactQueryDevtools } from 'react-query/devtools';
 
 const queryClient = new QueryClient();
 
 export const App = () => {
   return (
-    <ChakraProvider theme={theme}>
-      <QueryClientProvider client={queryClient}>
-        <Router basename='/'>
-          <UnauthorizedHandler />
-          <Routes>
-            <Route element={<AnimatedTransition />}>
-              <Route element={<RequireAuthRoute />}>
-                <Route path={Path.EXPLORER} element={<Explorer />} />
-                <Route path={Path.CALENDAR + '/:yearCourseId'} element={<Calendar />} />
-                <Route path={Path.DASHBOARD + '/:yearCourseId'} element={<Dashboard />} />
-                <Route path='*' element={<Navigate to={Path.EXPLORER} />} />
-              </Route>
+    <ParallaxProvider>
+      <ChakraProvider theme={theme}>
+        <QueryClientProvider client={queryClient}>
+          <ReactQueryDevtools initialIsOpen={false} />
+          <Router basename='/'>
+            <UnauthorizedHandler />
+            <Routes>
+              <Route element={<AnimatedTransition />}>
+                <Route path='*' element={<Navigate to={Path.LANDING_PAGE} />} />
 
-              <Route path={Path.LOGIN} element={<Login />} />
-              <Route path={Path.REGISTER} element={<Register />} />
-            </Route>
-          </Routes>
-        </Router>
-      </QueryClientProvider>
-    </ChakraProvider>
+                <Route path={Path.LANDING_PAGE} element={<LandingPage />} />
+                <Route path={Path.LOGIN} element={<Login />} />
+                <Route path={Path.REGISTER} element={<Register />} />
+
+                <Route element={<RequireAuthRoute />}>
+                  <Route path={Path.EXPLORER} element={<Explorer />} />
+                  <Route path={Path.CALENDAR + '/:yearCourseId'} element={<Calendar />} />
+                  <Route path={Path.DASHBOARD + '/:yearCourseId'} element={<Dashboard />} />
+                </Route>
+              </Route>
+            </Routes>
+          </Router>
+        </QueryClientProvider>
+      </ChakraProvider>
+    </ParallaxProvider>
   );
 };
