@@ -16,10 +16,10 @@ export const SubjectListItem = (props: Props) => {
     <>
       <SubjectEditModal isOpen={isEditModalOpen} onClose={onEditModalClose} subject={props.subject} />
       <SubjectDeleteModal isOpen={isDeleteModalOpen} onClose={onDeleteModalClose} subject={props.subject} />
-      <Flex py={2} borderRadius={5}>
+      <Flex py={2} borderRadius={5} alignItems={'center'}>
         <VStack spacing={0} alignItems={'baseline'}>
           <Text>{props.subject.shortName}</Text>
-          <Text fontSize={'xs'} color={'gray.400'}>
+          <Text fontSize={'sm'} color={'gray.400'}>
             {props.subject.name}
           </Text>
         </VStack>
