@@ -1,10 +1,11 @@
-import { AddIcon, ArrowBackIcon, HamburgerIcon, SettingsIcon } from '@chakra-ui/icons';
-import { Button, Flex, Heading, HStack, Spacer, Text, VStack } from '@chakra-ui/react';
 import { useNavigate, useParams } from 'react-router-dom';
-import { CalendarPanel } from '../calendar/CalendarPanel';
-import { EventsPanel } from '../components/calendar/EventsPanel';
+import { CalendarPanel } from '../components/calendar/days/CalendarPanel';
+import { EventsPanel } from '../components/calendar/events/EventsPanel';
 import useAuth from '../hooks/useAuth';
 import { Path } from '../other/Paths';
+import { MdDashboard } from 'react-icons/md';
+import { HiOutlineLogout, HiOutlineViewGrid } from 'react-icons/hi';
+import { Flex, Heading, Spacer, Button, Text } from '@chakra-ui/react';
 
 export const Calendar = () => {
   const navigate = useNavigate();
@@ -13,25 +14,25 @@ export const Calendar = () => {
   const { yearCourseId } = useParams();
 
   return (
-    <Flex p={4} m={0} h={['auto', '100vh']} direction={'column'} overflow={'hidden'}>
+    <Flex p={[2, 4]} m={0} h={['100vh']} direction={'column'} overflow={'hidden'}>
       <Flex m={0} p={0} mb={2}>
         <Heading>Kalendarz</Heading>
         <Spacer />
-        <Button onClick={() => navigate(Path.DASHBOARD + `/${yearCourseId}`)} leftIcon={<HamburgerIcon />}>
-          Panel zarządzania
+        <Button onClick={() => navigate(Path.DASHBOARD + `/${yearCourseId}`)} leftIcon={<MdDashboard />}>
+          <Text display={['none', 'unset']}>Panel zarządzania</Text>
         </Button>
-        <Button ml={3} onClick={() => navigate(Path.EXPLORER)} leftIcon={<ArrowBackIcon />}>
-          Wybór kierunku
+        <Button ml={3} onClick={() => navigate(Path.EXPLORER)} leftIcon={<HiOutlineViewGrid />}>
+          <Text display={['none', 'unset']}>Wybór kierunku</Text>
         </Button>
-        <Button ml={3} onClick={() => logout()}>
-          Wyloguj
+        <Button ml={3} onClick={() => logout()} leftIcon={<HiOutlineLogout />}>
+          <Text display={['none', 'unset']}>Wyloguj</Text>
         </Button>
       </Flex>
 
-      <HStack flexGrow={1} spacing={4}>
+      <Flex flexGrow={1} gap={4} direction={['column', 'row']} overflow={'hidden'} m={-10} p={10}>
         <CalendarPanel />
         <EventsPanel />
-      </HStack>
+      </Flex>
     </Flex>
   );
 };
