@@ -1,12 +1,24 @@
-import { Flex, Spacer, Box, Heading, Button, SimpleGrid, useDisclosure, Grid } from '@chakra-ui/react';
+import {
+  Flex,
+  Spacer,
+  Box,
+  Heading,
+  Button,
+  SimpleGrid,
+  useDisclosure,
+  Grid,
+  Center,
+  VStack,
+  HStack,
+  Text,
+} from '@chakra-ui/react';
 import useAuth from '../hooks/useAuth';
 import useYearCourses from '../hooks/useYearCourses';
-import { ColorModeSwitcher } from '../components/other/ColorModeSwitcher';
-
 import { YearCourseCard } from '../components/explorer/YearCourseCard';
 import { YearCourseCreateModal } from '../components/explorer/YearCourseCreateModal';
 import { AddIcon, PlusSquareIcon } from '@chakra-ui/icons';
 import { YearCourseType } from '../entities/YearCourse';
+import { HiOutlineLogout } from 'react-icons/hi';
 
 export const Explorer = () => {
   const { isOpen, onOpen, onClose } = useDisclosure();
@@ -14,25 +26,27 @@ export const Explorer = () => {
   const { logout } = useAuth();
 
   return (
-    <Box p={5}>
+    <Flex p={4} m={0} h={['100vh']} direction={'column'} overflow={'hidden'}>
       <YearCourseCreateModal isOpen={isOpen} onClose={onClose} />
       <Flex mb={5}>
-        <Heading>Eksplorator kierunków</Heading>
-        <Button ml={5} onClick={onOpen} rightIcon={<PlusSquareIcon />} variant={'outline'}>
-          Dodaj kierunek
-        </Button>
         <Spacer />
-        <ColorModeSwitcher />
-        <Button ml={3} onClick={() => logout()}>
+        <Button ml={3} onClick={() => logout()} leftIcon={<HiOutlineLogout />}>
           Wyloguj
         </Button>
       </Flex>
-      <SimpleGrid columns={[1, 2, 3, 4, 5]} gap={6}>
-        {yearCourseQuery.data &&
-          yearCourseQuery.data.map((yearCourse: YearCourseType) => (
-            <YearCourseCard key={yearCourse.id} yearCourse={yearCourse} />
-          ))}
-      </SimpleGrid>
-    </Box>
+      <Center flexGrow={1} flexDirection={'column'}>
+        <HStack gap={6} w={'100%'} mt={-10}>
+          <Spacer />
+          {yearCourseQuery.data &&
+            yearCourseQuery.data.map((yearCourse: YearCourseType) => (
+              <YearCourseCard key={yearCourse.id} yearCourse={yearCourse} />
+            ))}
+          <Button ml={5} onClick={onOpen} rightIcon={<PlusSquareIcon />} variant={'outline'} w={'20%'} h={'100%'}>
+            Dodaj kierunek
+          </Button>
+          <Spacer />
+        </HStack>
+      </Center>
+    </Flex>
   );
 };
