@@ -1,10 +1,12 @@
 import create from 'zustand';
 import { devtools, persist } from 'zustand/middleware';
 import { createAuthSlice } from './authSlice';
-import axios from "axios";
+import axios from 'axios';
+import { createCalendarSlice } from './calendarSlice';
 
 const store = (set: any) => ({
   ...createAuthSlice(set),
+  ...createCalendarSlice(set),
 });
 
 const useStore = create(
